+++ src/Pixeval/Pages/Capability/MyPixivUsersPage.xaml
@@ -1,5 +1,5 @@
 <controls:EnhancedPage
-    x:Class="Pixeval.Pages.MyPixivPage"
+    x:Class="Pixeval.Pages.MyPixivUsersPage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:controls="using:Pixeval.Controls"
