+++ src/Pixeval/Pages/Capability/RecommendUsersPage.xaml.cs
@@ -1,4 +1,3 @@
-using System.Linq;
 using Microsoft.UI.Xaml.Controls;
 using Microsoft.UI.Xaml.Navigation;
 using Pixeval.Misc;
@@ -9,14 +8,9 @@ public sealed partial class RecommendUsersPage : IScrollViewProvider
 {
     public RecommendUsersPage() => InitializeComponent();
 
-    public override async void OnPageActivated(NavigationEventArgs e)
+    public override void OnPageActivated(NavigationEventArgs e)
     {
-        if (e.Parameter is not long userId)
-            userId = App.AppViewModel.PixivUid;
-
-        var engine = App.AppViewModel.MakoClient.Computed((await App.AppViewModel.MakoClient.RelatedUserAsync(userId))
-            .Users.ToAsyncEnumerable());
-        IllustratorView.ViewModel.ResetEngine(engine);
+        IllustratorView.ViewModel.ResetEngine(App.AppViewModel.MakoClient.RecommendIllustrators());
     }
 
     public ScrollView ScrollView => IllustratorView.ScrollView;
