+++ src/Pixeval/Pages/Capability/RelatedUsersPage.xaml
@@ -1,0 +1,11 @@
+<controls:EnhancedPage
+    x:Class="Pixeval.Pages.Capability.RelatedUsersPage"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:controls="using:Pixeval.Controls"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:local="using:Pixeval.Pages.Capability"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    mc:Ignorable="d">
+    <controls:IllustratorView x:Name="IllustratorView" />
+</controls:EnhancedPage>
