+++ src/Pixeval/Strings/en-us/MainPage.resjson
@@ -1,6 +1,6 @@
 {
     "BookmarksTab/Content": "Bookmark",
-    "FeedsTab/Content": "Feed",
+    "RecommendUsersTab/Content": "Recommend Users",
     "FollowingsTab/Content": "Following",
     "HistoriesTab/Content": "History",
     "TagsTab/Content": "My Tags",
