+++ src/Pixeval/Strings/ru-ru/EntryViewerPage.resjson
@@ -28,7 +28,6 @@
   
   "FollowingsNavigationViewItem/Content": "Подписки",
   "MyPixivUserNavigationViewItem/Content": "Друзья в Pixiv",
-  "RecommendUserNavigationViewItem/Content": "Рекомендуемые пользователи",
   "FollowingUserItem/Title": "Подписки",
   "MyPixivUserItem/Title": "Друзья в Pixiv",
   "IllustrationCountItem/Title": "Количество иллюстраций",
