+++ src/Pixeval/Strings/ru-ru/MainPage.resjson
@@ -1,6 +1,5 @@
 {
     "BookmarksTab/Content": "Моё избранное",
-    "FeedsTab/Content": "Лента подписок",
     "FollowingsTab/Content": "Мои подписки",
     "HistoriesTab/Content": "История просмотров",
     "TagsTab/Content": "Управление тегами",
