+++ src/Pixeval/Strings/zh-cn/MainPage.resjson
@@ -1,6 +1,6 @@
 {
     "BookmarksTab/Content": "我的收藏",
-    "FeedsTab/Content": "关注动态",
+    "RecommendUsersTab/Content": "推荐作者",
     "FollowingsTab/Content": "我的关注",
     "HistoriesTab/Content": "浏览历史",
     "TagsTab/Content": "标签管理",
